requirements.txt
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.17.3
transformers==4.32.0
PyPDF2==3.0.1
pydantic==2.5.0
PyYAML==6.0.1
numpy==1.24.3
python-dateutil==2.8.2
pytest==7.4.3
torch>=1.11.0
tokenizers==0.13.3
reportlab
google-generativeai
pinecone
PyMuPDF>=1.23.0        # Best for PDF and DOCX (RECOMMENDED)
PyPDF2>=3.0.0
python-docx>=0.8.11
chromadb>=0.4.0
gradio>=4.0.0

